+++ api/src/models/young.js
@@ -852,6 +852,12 @@ const Schema = new mongoose.Schema({
       description: "Le parent 1 s'est identifié via France Connect",
     },
   },
+  parent1Inscription2023Token: {
+    type: String,
+    documentation: {
+      description: "Token d'inscription 2023 du parent 1",
+    },
+  },
 
   parent2Status: {
     type: String,
@@ -942,6 +948,12 @@ const Schema = new mongoose.Schema({
     default: "false",
     documentation: {
       description: "Le parent 2 s'est identifié via France Connect",
+    },
+  },
+  parent2Inscription2023Token: {
+    type: String,
+    documentation: {
+      description: "Token d'inscription 2023 du parent 1",
     },
   },
 
@@ -1587,6 +1599,8 @@ Schema.plugin(
       "invitationExpires",
       "phase3Token",
       "loginAttempts",
+      "parent1Inscription2023Token",
+      "parent2Inscription2023Token",
     ],
   }),
   MODELNAME,