+++ app/src/scenes/inscription2023/components/GhostButton.js
@@ -1,8 +1,14 @@
 import React from "react";
 
-const GhostButton = ({ onClick, name, type = undefined }) => {
+const GhostButton = ({ onClick, name, type = undefined, disabled = false }) => {
   return (
-    <button className="mt-2 mb-6 max-w-md bg-white w-full py-2 border !border-[#000091] shadow-sm text-[#000091] font-medium" type={type} onClick={onClick}>
+    <button
+      disabled={disabled}
+      className={`mt-2 mb-6 max-w-md w-full py-2 font-medium ${
+        disabled ? "bg-[#E5E5E5] text-[#929292] cursor-not-allowed" : "bg-white border !border-[#000091] shadow-sm text-[#000091]"
+      }`}
+      type={type}
+      onClick={onClick}>
       {name}
     </button>
   );