+++ app/src/scenes/inscription2023/components/Select.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { BsCheck2 } from "react-icons/bs";
 import { IoMdClose } from "react-icons/io";
+import ErrorMessage from "./ErrorMessage";
 import ChevronDown from "../../../assets/icons/ChevronDown";
 
 export default function Select({ label, options, value, placeholder = "Sélectionner une option", Icon = null, alignItems = "left", onChange, error = "" }) {
@@ -29,7 +30,7 @@ export default function Select({ label, options, value, placeholder = "Sélectio
   };
 
   return (
-    <div className="my-2 text-[#161616]" style={{ fontFamily: "Marianne" }} ref={ref}>
+    <div className="mt-2 mb-6 text-[#161616]" style={{ fontFamily: "Marianne" }} ref={ref}>
       <label className="my-2">{label}</label>
       <div className="relative">
         {/* select item */}
@@ -68,7 +69,7 @@ export default function Select({ label, options, value, placeholder = "Sélectio
           ))}
         </div>
       </div>
-      <div className="h-4 text-[#CE0500] text-sm">{error}</div>
+      <ErrorMessage>{error}</ErrorMessage>
     </div>
   );
 }